+++ src/generators/Ripple.hpp
@@ -12,22 +12,22 @@
 class Ripple: public Generator {
 
 public:
-    static float get( int& sync, float& ripple, float& lastR ) {
+    static double get( int& sync, double& ripple, double& lastR ) {
         int line = sync % Config::SCREEN_HEIGHT;
-        float screenpos = (float) line / Config::SCREEN_HEIGHT;
-        float vt = abs(sin(M_PI *  screenpos));
-        float ret = vt > lastR? vt: lastR;
-        float correct = 1 - (10e-4 * ripple);
+        double screenpos = (double) line / Config::SCREEN_HEIGHT;
+        double vt = abs(sin(M_PI *  screenpos));
+        double ret = vt > lastR? vt: lastR;
+        double correct = 1 - (10e-4 * ripple);
         lastR = ret * correct;
         return ret ;
     }
 
-    float get(int& sync, float& ripple) {
+    double get(int& sync, double& ripple) {
         return get( sync, ripple, hold_state );
     }
 
 protected:
-    float hold_state = 0;
+    double hold_state = 0;
 };
 
 #endif //SDL_CRT_FILTER_RIPPLE_HPP
